+++ src/main/java/io/tpd/superheroes/MvcTestsApplication.java
@@ -1,4 +1,4 @@
-package es.macero.dev;
+package io.tpd.superheroes;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
