+++ src/main/java/io/tpd/superheroes/controller/SuperHeroController.java
@@ -1,7 +1,7 @@
-package es.macero.dev.controller;
+package io.tpd.superheroes.controller;
 
-import es.macero.dev.domain.SuperHero;
-import es.macero.dev.repository.SuperHeroRepository;
+import io.tpd.superheroes.domain.SuperHero;
+import io.tpd.superheroes.repository.SuperHeroRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
