+++ src/test/java/io/tpd/superheroes/controller/SuperHeroControllerMockMvcStandaloneTest.java
@@ -1,9 +1,9 @@
-package es.macero.dev.controller;
+package io.tpd.superheroes.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import es.macero.dev.domain.SuperHero;
-import es.macero.dev.exceptions.NonExistingHeroException;
-import es.macero.dev.repository.SuperHeroRepository;
+import io.tpd.superheroes.domain.SuperHero;
+import io.tpd.superheroes.exceptions.NonExistingHeroException;
+import io.tpd.superheroes.repository.SuperHeroRepository;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
