+++ src/test/java/io/tpd/superheroes/controller/SuperHeroControllerMockMvcWithContextTest.java
@@ -1,15 +1,14 @@
-package es.macero.dev.controller;
+package io.tpd.superheroes.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import es.macero.dev.domain.SuperHero;
-import es.macero.dev.exceptions.NonExistingHeroException;
-import es.macero.dev.repository.SuperHeroRepository;
+import io.tpd.superheroes.domain.SuperHero;
+import io.tpd.superheroes.exceptions.NonExistingHeroException;
+import io.tpd.superheroes.repository.SuperHeroRepository;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.json.JacksonTester;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.HttpStatus;
@@ -26,15 +25,13 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 
 /**
- * This class demonstrates how to test a controller using Spring Boot Test
- * with a MOCK web environment, which makes it similar to just using @WebMvcTest
+ * This class demonstrates how to test a controller using MockMVC loading a Test Context
  *
  * @author moises.macero
  */
 @RunWith(SpringRunner.class)
-@SpringBootTest
-@AutoConfigureMockMvc
-public class SuperHeroControllerSpringBootMockTest {
+@WebMvcTest(SuperHeroController.class)
+public class SuperHeroControllerMockMvcWithContextTest {
 
     @Autowired
     private MockMvc mvc;
