+++ src/test/java/io/tpd/superheroes/controller/SuperHeroControllerSpringBootMockTest.java
@@ -1,14 +1,15 @@
-package es.macero.dev.controller;
+package io.tpd.superheroes.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import es.macero.dev.domain.SuperHero;
-import es.macero.dev.exceptions.NonExistingHeroException;
-import es.macero.dev.repository.SuperHeroRepository;
+import io.tpd.superheroes.domain.SuperHero;
+import io.tpd.superheroes.exceptions.NonExistingHeroException;
+import io.tpd.superheroes.repository.SuperHeroRepository;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.json.JacksonTester;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.HttpStatus;
@@ -25,13 +26,15 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 
 /**
- * This class demonstrates how to test a controller using MockMVC loading a Test Context
+ * This class demonstrates how to test a controller using Spring Boot Test
+ * with a MOCK web environment, which makes it similar to just using @WebMvcTest
  *
  * @author moises.macero
  */
 @RunWith(SpringRunner.class)
-@WebMvcTest(SuperHeroController.class)
-public class SuperHeroControllerMockMvcWithContextTest {
+@SpringBootTest
+@AutoConfigureMockMvc
+public class SuperHeroControllerSpringBootMockTest {
 
     @Autowired
     private MockMvc mvc;
